feat: port util.js tap test to node-test
